refactor: rename JuejinLogo to JuejinIcon and remove RiddleIcon
